backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai>=2.0.0
requests==2.31.0
python-docx==1.1.0
gunicorn==21.2.0

